requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
astroid==2.2.5
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
colorlog==4.0.2
cryptography==2.6.1
Django==2.2
djangorestframework==3.9.2
Flask==1.0.2
Flask-KVSession==0.6.2
Flask-RESTful==0.3.7
Flask-WTF==0.14.2
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
marshmallow==3.0.0rc5
mccabe==0.6.1
mysqlclient==1.4.2.post1
peewee==3.9.3
pycparser==2.19
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
PyMySQL==0.9.3
python-dotenv==0.10.1
pytz==2018.9
simplekv==0.12.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
Werkzeug==0.15.1
wrapt==1.11.1
WTForms==2.2.1
